Separate Offer and Cart functions
